refactor: enforce strict controller → service → repository flow
